+++ packages/target-ethers-v5-test/types/factories/v0.6.4/DataTypesView__factory.ts
@@ -214,7 +214,7 @@ const _abi = [
     stateMutability: "view",
     type: "function",
   },
-];
+] as const;
 
 export class DataTypesView__factory {
   static readonly abi = _abi;