+++ packages/target-ethers-v5-test/types/factories/v0.6.4/LibraryConsumer__factory.ts
@@ -29,7 +29,7 @@ const _abi = [
     stateMutability: "pure",
     type: "function",
   },
-];
+] as const;
 
 export class LibraryConsumer__factory {
   static readonly abi = _abi;