+++ packages/target-ethers-v5-test/types/factories/v0.6.4/Payable/PayableFactory__factory.ts
@@ -23,7 +23,7 @@ const _abi = [
     stateMutability: "nonpayable",
     type: "function",
   },
-];
+] as const;
 
 export class PayableFactory__factory {
   static readonly abi = _abi;