+++ packages/target-ethers-v5-test/types/factories/v0.8.9/KingOfTheHill/KingOfTheHill__factory.ts
@@ -88,7 +88,7 @@ const _abi = [
     stateMutability: "nonpayable",
     type: "function",
   },
-];
+] as const;
 
 export class KingOfTheHill__factory {
   static readonly abi = _abi;