+++ packages/target-ethers-v5-test/types/factories/v0.8.9/Rarity/ERC721__factory.ts
@@ -273,7 +273,7 @@ const _abi = [
     stateMutability: "nonpayable",
     type: "function",
   },
-];
+] as const;
 
 export class ERC721__factory {
   static readonly abi = _abi;