+++ packages/target-ethers-v5-test/types/factories/v0.8.9/Rarity/IERC721Receiver__factory.ts
@@ -44,7 +44,7 @@ const _abi = [
     stateMutability: "nonpayable",
     type: "function",
   },
-];
+] as const;
 
 export class IERC721Receiver__factory {
   static readonly abi = _abi;