+++ packages/target-ethers-v5-test/types/factories/v0.8.9/SimpleToken__factory.ts
@@ -28,7 +28,7 @@ const _abi = [
     stateMutability: "nonpayable",
     type: "function",
   },
-];
+] as const;
 
 export class SimpleToken__factory {
   static readonly abi = _abi;