+++ packages/target-ethers-v6/test/generation.test.ts
@@ -1,0 +1,97 @@
+/* eslint-disable import/no-extraneous-dependencies */
+import { expect } from 'earljs'
+import { CodegenConfig, Contract, EventDeclaration, parse, RawAbiDefinition } from 'typechain'
+
+import { codegenContractFactory } from '../src/codegen'
+import { generateEventTypeExport } from '../src/codegen/events'
+
+const DEFAULT_FLAGS: CodegenConfig = {
+  alwaysGenerateOverloads: false,
+  discriminateTypes: false,
+  tsNocheck: false,
+  environment: undefined,
+}
+
+describe('Ethers generation edge cases', () => {
+  const emptyContract: Contract = {
+    name: 'TestContract',
+    rawName: 'TestContract',
+    functions: {},
+    events: {},
+    structs: {},
+    constructor: [{ name: 'constructor', inputs: [], outputs: [], stateMutability: 'nonpayable' }],
+    path: [],
+  }
+
+  it('should generate simple factory when no bytecode available', () => {
+    expect(codegenContractFactory(DEFAULT_FLAGS, emptyContract, 'abi', undefined)).toEqual(
+      expect.stringMatching(/export class TestContract__factory \{/),
+    )
+  })
+
+  it('should work when bytecode has linkReferences', () => {
+    const source = codegenContractFactory(DEFAULT_FLAGS, emptyContract, 'abi', {
+      bytecode: '{{BYTECODE}}',
+      linkReferences: [{ reference: '{{REFERENCE}}' }],
+    })
+
+    expect(source).toEqual(expect.stringMatching(/export class TestContract__factory extends ContractFactory \{/))
+    expect(source).toEqual(expect.stringMatching(/static linkBytecode\(/))
+  })
+})
+
+describe(generateEventTypeExport.name, () => {
+  it('generates proper argument names even if arguments are not named in Solidity', () => {
+    const event: EventDeclaration = {
+      name: 'UpdateFrequencySet',
+      isAnonymous: false,
+      inputs: [
+        {
+          name: undefined,
+          isIndexed: false,
+          type: {
+            type: 'array',
+            itemType: { type: 'address', originalType: 'address' },
+            originalType: 'address[]',
+          },
+        },
+        {
+          name: undefined,
+          isIndexed: false,
+          type: {
+            type: 'array',
+            itemType: { type: 'uinteger', bits: 256, originalType: 'uint256' },
+            originalType: 'uint256[]',
+          },
+        },
+      ],
+    }
+
+    const actual = generateEventTypeExport(event, false)
+    expect(actual).toEqual(expect.stringMatching('arg0: string[]'))
+    expect(actual).toEqual(expect.stringMatching('arg1: bigint[]'))
+  })
+
+  it('generates argument names from event field names in the ABI', () => {
+    const abi = [
+      {
+        anonymous: false,
+        inputs: [
+          { indexed: true, internalType: 'address', name: 'owner', type: 'address' },
+          { indexed: true, internalType: 'address', name: 'approved', type: 'address' },
+          { indexed: true, internalType: 'uint256', name: 'tokenId', type: 'uint256' },
+        ],
+        name: 'Approval',
+        type: 'event',
+      },
+    ] as any as RawAbiDefinition[]
+
+    const contract = parse(abi, 'Rarity')
+
+    const actual = generateEventTypeExport(Object.values(contract.events)[0][0], false)
+
+    expect(actual).toEqual(expect.stringMatching('owner: string'))
+    expect(actual).toEqual(expect.stringMatching('approved: string'))
+    expect(actual).toEqual(expect.stringMatching('tokenId: BigNumberish'))
+  })
+})