+++ packages/test-e2e/CHANGELOG.md
@@ -1,5 +1,20 @@
 # test-e2e
 
+## 9.9.13
+
+### Patch Changes
+
+- Updated dependencies [31e6957]
+  - @typechain/ethers-v5@10.2.0
+
+## 9.9.12
+
+### Patch Changes
+
+- Updated dependencies [bbc9656]
+  - typechain@8.1.1
+  - @typechain/ethers-v5@10.1.1
+
 ## 9.9.11
 
 ### Patch Changes